Move getCompaniesOptions and getContactsOptions methods from LeadController
